move project data to json
