+++ app/utils/constants.ts
@@ -18,3 +18,5 @@ export const NAV_ROUTES = [
 ]
 
 export const WEB_FORM_KEY = '5602ac20-bcd3-48da-bd9e-51ddde3b01df'
+
+export const PROJECTS_URL = '/projects'
